Move ConversationHistoryManager in chat.ts for chat functionality consolidation. Remove extraneous grammar and clean up index.ts exports. Update imports in lm.ts and vlm.ts to reflect new structure.
